+++ shared/src/config.rs
@@ -1,5 +1,6 @@
 use {
     base64::{engine::general_purpose::STANDARD as base64_engine, Engine},
+    rustls::pki_types::{CertificateDer, PrivateKeyDer, PrivatePkcs8KeyDer},
     serde::{
         de::{self, Deserializer},
         Deserialize,
@@ -8,8 +9,9 @@ use {
     std::{
         collections::HashSet,
         fmt::Display,
-        io,
+        fs, io,
         net::{IpAddr, Ipv4Addr, SocketAddr},
+        path::PathBuf,
         str::FromStr,
         sync::atomic::{AtomicU64, Ordering},
     },
@@ -17,26 +19,16 @@ use {
 };
 
 #[derive(Debug, Clone, Default, Deserialize)]
-#[serde(default, deny_unknown_fields)]
+#[serde(deny_unknown_fields, default)]
 pub struct ConfigTokio {
     /// Number of worker threads in Tokio runtime
     pub worker_threads: Option<usize>,
     /// Threads affinity
-    #[serde(deserialize_with = "ConfigTokio::deserialize_affinity")]
+    #[serde(deserialize_with = "deserialize_affinity")]
     pub affinity: Option<Vec<usize>>,
 }
 
 impl ConfigTokio {
-    pub fn deserialize_affinity<'de, D>(deserializer: D) -> Result<Option<Vec<usize>>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        match Option::<&str>::deserialize(deserializer)? {
-            Some(taskset) => parse_taskset(taskset).map(Some).map_err(de::Error::custom),
-            None => Ok(None),
-        }
-    }
-
     pub fn build_runtime<T>(self, thread_name_prefix: T) -> io::Result<tokio::runtime::Runtime>
     where
         T: AsRef<str> + Send + Sync + 'static,
@@ -61,54 +53,8 @@ impl ConfigTokio {
     }
 }
 
-pub fn parse_taskset(taskset: &str) -> Result<Vec<usize>, String> {
-    let mut set = HashSet::new();
-    for taskset2 in taskset.split(',') {
-        match taskset2.split_once('-') {
-            Some((start, end)) => {
-                let start: usize = start
-                    .parse()
-                    .map_err(|_error| format!("failed to parse {start:?} from {taskset:?}"))?;
-                let end: usize = end
-                    .parse()
-                    .map_err(|_error| format!("failed to parse {end:?} from {taskset:?}"))?;
-                if start > end {
-                    return Err(format!("invalid interval {taskset2:?} in {taskset:?}"));
-                }
-                for idx in start..=end {
-                    set.insert(idx);
-                }
-            }
-            None => {
-                set.insert(
-                    taskset2.parse().map_err(|_error| {
-                        format!("failed to parse {taskset2:?} from {taskset:?}")
-                    })?,
-                );
-            }
-        }
-    }
-
-    let mut vec = set.into_iter().collect::<Vec<usize>>();
-    vec.sort();
-
-    if let Some(set_max_index) = vec.last().copied() {
-        let max_index = affinity::get_thread_affinity()
-            .map_err(|_err| "failed to get affinity".to_owned())?
-            .into_iter()
-            .max()
-            .unwrap_or(0);
-
-        if set_max_index > max_index {
-            return Err(format!("core index must be in the range [0, {max_index}]"));
-        }
-    }
-
-    Ok(vec)
-}
-
 #[derive(Debug, Clone, Copy, Deserialize)]
-#[serde(default, deny_unknown_fields)]
+#[serde(deny_unknown_fields, default)]
 pub struct ConfigPrometheus {
     /// Endpoint of Prometheus service
     pub endpoint: SocketAddr,
@@ -226,4 +172,149 @@ where
     let sig: Option<&str> = Deserialize::deserialize(deserializer)?;
     sig.map(|sig| sig.parse().map_err(de::Error::custom))
         .transpose()
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(deny_unknown_fields, untagged)]
+enum RustlsServerConfigSignedSelfSigned<'a> {
+    Signed { cert: &'a str, key: &'a str },
+    SelfSigned { self_signed_alt_names: Vec<String> },
+}
+
+impl<'a> RustlsServerConfigSignedSelfSigned<'a> {
+    fn parse<D>(self) -> Result<rustls::ServerConfig, D::Error>
+    where
+        D: Deserializer<'a>,
+    {
+        let (certs, key) = match self {
+            Self::Signed { cert, key } => {
+                let cert_path = PathBuf::from(cert);
+                let cert_bytes = fs::read(&cert_path).map_err(|error| {
+                    de::Error::custom(format!("failed to read cert {cert_path:?}: {error:?}"))
+                })?;
+                let cert_chain = if cert_path.extension().is_some_and(|x| x == "der") {
+                    vec![CertificateDer::from(cert_bytes)]
+                } else {
+                    rustls_pemfile::certs(&mut &*cert_bytes)
+                        .collect::<Result<_, _>>()
+                        .map_err(|error| {
+                            de::Error::custom(format!("invalid PEM-encoded certificate: {error:?}"))
+                        })?
+                };
+
+                let key_path = PathBuf::from(key);
+                let key_bytes = fs::read(&key_path).map_err(|error| {
+                    de::Error::custom(format!("failed to read key {key_path:?}: {error:?}"))
+                })?;
+                let key = if key_path.extension().is_some_and(|x| x == "der") {
+                    PrivateKeyDer::Pkcs8(PrivatePkcs8KeyDer::from(key_bytes))
+                } else {
+                    rustls_pemfile::private_key(&mut &*key_bytes)
+                        .map_err(|error| {
+                            de::Error::custom(format!("malformed PKCS #1 private key: {error:?}"))
+                        })?
+                        .ok_or_else(|| de::Error::custom("no private keys found"))?
+                };
+
+                (cert_chain, key)
+            }
+            Self::SelfSigned {
+                self_signed_alt_names,
+            } => {
+                let cert =
+                    rcgen::generate_simple_self_signed(self_signed_alt_names).map_err(|error| {
+                        de::Error::custom(format!("failed to generate self-signed cert: {error:?}"))
+                    })?;
+                let cert_der = CertificateDer::from(cert.cert);
+                let priv_key = PrivatePkcs8KeyDer::from(cert.key_pair.serialize_der());
+                (vec![cert_der], priv_key.into())
+            }
+        };
+
+        rustls::ServerConfig::builder()
+            .with_no_client_auth()
+            .with_single_cert(certs, key)
+            .map_err(|error| de::Error::custom(format!("failed to use cert: {error:?}")))
+    }
+}
+
+pub fn deserialize_maybe_rustls_server_config<'de, D>(
+    deserializer: D,
+) -> Result<Option<rustls::ServerConfig>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let config: Option<RustlsServerConfigSignedSelfSigned> =
+        Deserialize::deserialize(deserializer)?;
+    if let Some(config) = config {
+        config.parse::<D>().map(Some)
+    } else {
+        Ok(None)
+    }
+}
+
+pub fn deserialize_rustls_server_config<'de, D>(
+    deserializer: D,
+) -> Result<rustls::ServerConfig, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let config: RustlsServerConfigSignedSelfSigned = Deserialize::deserialize(deserializer)?;
+    config.parse::<D>()
+}
+
+pub fn deserialize_affinity<'de, D>(deserializer: D) -> Result<Option<Vec<usize>>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    match Option::<&str>::deserialize(deserializer)? {
+        Some(taskset) => parse_taskset(taskset).map(Some).map_err(de::Error::custom),
+        None => Ok(None),
+    }
+}
+
+pub fn parse_taskset(taskset: &str) -> Result<Vec<usize>, String> {
+    let mut set = HashSet::new();
+    for taskset2 in taskset.split(',') {
+        match taskset2.split_once('-') {
+            Some((start, end)) => {
+                let start: usize = start
+                    .parse()
+                    .map_err(|_error| format!("failed to parse {start:?} from {taskset:?}"))?;
+                let end: usize = end
+                    .parse()
+                    .map_err(|_error| format!("failed to parse {end:?} from {taskset:?}"))?;
+                if start > end {
+                    return Err(format!("invalid interval {taskset2:?} in {taskset:?}"));
+                }
+                for idx in start..=end {
+                    set.insert(idx);
+                }
+            }
+            None => {
+                set.insert(
+                    taskset2.parse().map_err(|_error| {
+                        format!("failed to parse {taskset2:?} from {taskset:?}")
+                    })?,
+                );
+            }
+        }
+    }
+
+    let mut vec = set.into_iter().collect::<Vec<usize>>();
+    vec.sort();
+
+    if let Some(set_max_index) = vec.last().copied() {
+        let max_index = affinity::get_thread_affinity()
+            .map_err(|_err| "failed to get affinity".to_owned())?
+            .into_iter()
+            .max()
+            .unwrap_or(0);
+
+        if set_max_index > max_index {
+            return Err(format!("core index must be in the range [0, {max_index}]"));
+        }
+    }
+
+    Ok(vec)
 }