+++ shared/src/transports/quic.rs
@@ -1,6 +1,6 @@
 use {
     crate::{
-        config::deserialize_x_token_set,
+        config::{deserialize_rustls_server_config, deserialize_x_token_set},
         shutdown::Shutdown,
         transports::{RecvError, RecvItem, RecvStream, Subscribe, SubscribeError, WriteVectored},
     },
@@ -17,19 +17,13 @@ use {
         QuicSubscribeClose, QuicSubscribeCloseError, QuicSubscribeRequest, QuicSubscribeResponse,
         QuicSubscribeResponseError,
     },
-    rustls::pki_types::{CertificateDer, PrivateKeyDer, PrivatePkcs8KeyDer},
-    serde::{
-        de::{self, Deserializer},
-        Deserialize,
-    },
+    serde::Deserialize,
     std::{
         borrow::Cow,
         collections::{BTreeSet, HashSet, VecDeque},
-        fs,
         future::Future,
         io::{self, IoSlice},
         net::{IpAddr, Ipv4Addr, SocketAddr},
-        path::PathBuf,
         sync::Arc,
     },
     thiserror::Error,
@@ -45,7 +39,7 @@ use {
 pub struct ConfigQuicServer {
     #[serde(default = "ConfigQuicServer::default_endpoint")]
     pub endpoint: SocketAddr,
-    #[serde(deserialize_with = "ConfigQuicServer::deserialize_tls_config")]
+    #[serde(deserialize_with = "deserialize_rustls_server_config")]
     pub tls_config: rustls::ServerConfig,
     /// Value in ms
     #[serde(default = "ConfigQuicServer::default_expected_rtt")]
@@ -69,68 +63,6 @@ pub struct ConfigQuicServer {
 impl ConfigQuicServer {
     pub const fn default_endpoint() -> SocketAddr {
         SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 10100)
-    }
-
-    fn deserialize_tls_config<'de, D>(deserializer: D) -> Result<rustls::ServerConfig, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Debug, Deserialize)]
-        #[serde(deny_unknown_fields, untagged)]
-        enum Config<'a> {
-            Signed { cert: &'a str, key: &'a str },
-            SelfSigned { self_signed_alt_names: Vec<String> },
-        }
-
-        let (certs, key) = match Config::deserialize(deserializer)? {
-            Config::Signed { cert, key } => {
-                let cert_path = PathBuf::from(cert);
-                let cert_bytes = fs::read(&cert_path).map_err(|error| {
-                    de::Error::custom(format!("failed to read cert {cert_path:?}: {error:?}"))
-                })?;
-                let cert_chain = if cert_path.extension().is_some_and(|x| x == "der") {
-                    vec![CertificateDer::from(cert_bytes)]
-                } else {
-                    rustls_pemfile::certs(&mut &*cert_bytes)
-                        .collect::<Result<_, _>>()
-                        .map_err(|error| {
-                            de::Error::custom(format!("invalid PEM-encoded certificate: {error:?}"))
-                        })?
-                };
-
-                let key_path = PathBuf::from(key);
-                let key_bytes = fs::read(&key_path).map_err(|error| {
-                    de::Error::custom(format!("failed to read key {key_path:?}: {error:?}"))
-                })?;
-                let key = if key_path.extension().is_some_and(|x| x == "der") {
-                    PrivateKeyDer::Pkcs8(PrivatePkcs8KeyDer::from(key_bytes))
-                } else {
-                    rustls_pemfile::private_key(&mut &*key_bytes)
-                        .map_err(|error| {
-                            de::Error::custom(format!("malformed PKCS #1 private key: {error:?}"))
-                        })?
-                        .ok_or_else(|| de::Error::custom("no private keys found"))?
-                };
-
-                (cert_chain, key)
-            }
-            Config::SelfSigned {
-                self_signed_alt_names,
-            } => {
-                let cert =
-                    rcgen::generate_simple_self_signed(self_signed_alt_names).map_err(|error| {
-                        de::Error::custom(format!("failed to generate self-signed cert: {error:?}"))
-                    })?;
-                let cert_der = CertificateDer::from(cert.cert);
-                let priv_key = PrivatePkcs8KeyDer::from(cert.key_pair.serialize_der());
-                (vec![cert_der], priv_key.into())
-            }
-        };
-
-        rustls::ServerConfig::builder()
-            .with_no_client_auth()
-            .with_single_cert(certs, key)
-            .map_err(|error| de::Error::custom(format!("failed to use cert: {error:?}")))
     }
 
     const fn default_expected_rtt() -> u32 {