+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 - richat: remove extra lock on clients queue ([#49](https://github.com/lamports-dev/richat/pull/49))
 - plugin-agave: set `nodelay` correctly for Tcp ([#53](https://github.com/lamports-dev/richat/pull/53))
+- richat: add minimal sleep ([#54](https://github.com/lamports-dev/richat/pull/54))
 
 ### Features